[CPU:Bugfix] int64 audit of byte-offset arithmetic for > 2 GB intermediate tensors
3D segmentation models (nnUNet-style ResEnc-UNet, e.g. SIAM v0.3 with
input shape (1, 1, 256, 256, 192), opset 17) produce intermediate
tensors larger than 2 GB during the Conv3DTurn2D rewrite. With those
sizes a number of `int * int * int` byte-offset and accumulator
computations across the CPU runtime silently overflow int32, leading
to runtime SIGSEGV (out-of-bounds memmove inside _blit) or, more
insidiously, wrong-but-not-crashing output (silent fp16-level drift
to ~26 % argmax agreement with the reference).

This commit promotes the affected accumulators/offsets to size_t /
int64_t across 10 hot paths. Every visible site that does
`<int> * <int32_t-stride> * <int-bytes>` for byte-offset arithmetic on
SIAM-scale tensors is fixed; small/inner kernels that operate on
elements (not bytes) are unchanged.

Patches:

  include/MNN/Tensor.hpp
    * Tensor::elementSize() now returns size_t (was int). The primary
      bug -- when usize() exceeds 2^31 bytes, the cast to int wraps
      negative and downstream buffer-size math (memcpy lengths,
      allocator requests) silently truncates. Fixing this unblocks
      most of the cascade.

  source/shape/ShapeReshape.cpp
    * totalSizeInput / totalSizeOutput accumulators promoted to
      int64_t. Without this, Reshape's `-1` dim computation produces
      negative values for SIAM's late decoder Convs and the op fails
      shape inference with the now-recognizable "Reshape error:
      -1879048192 -> -1877999616" log line (= 0x90000000 / 0x90100000
      = int32 wrap of 192 * 256 * 256 * 192).

  source/backend/cpu/CPURaster.cpp
    * `_blit` byte-offset arithmetic (z*stride[0]*bytes, etc.) cast to
      size_t before pointer addition. Same fix in `executeFaster`
      lambda and in the four `_blit` call sites elsewhere in the file.
    * Defensive workaround: demote any Raster op with a per-stride
      byte step > 16 MB to single-thread. SIAM-scale multi-thread
      execution still segfaults in CPURaster::onResize::lambda#4 even
      with the per-thread offset arithmetic patched (root cause not
      yet isolated; flag for follow-up).

  source/backend/cpu/CPUTensorConvert.cpp
    * v * pack * bitLength * area / v * channel * bitLength * area
      byte offsets in batched NC4HW4 conversions cast to size_t.

  source/backend/cpu/CPULayerNorm.cpp
    * tId * mInnerSize * bytes overflowed at tId >= 43 for SIAM's
      LayerNorm-over-(D, H, W) (mInnerSize ~ 12.5M, bytes = 4 -> ~ 50
      MB per tId step). Cast to size_t; same for the buf->alloc()
      size in onResize.

  source/backend/cpu/CPUScale.cpp
    * depthStride * i * bytes overflowed at i >= 6 for SIAM's Scale-
      after-LayerNorm (depthStride ~ 100M, bytes = 4). Hoisted to a
      size_t byteOffset local.

  source/backend/cpu/CPUDeconvolution.cpp
    * outputSize = batch * src_w * src_h * ocC4 * pack * bytes
      promoted to size_t. mExtraOutput allocation, memset, and the
      inner per-channel pointer offsets in both onResize lambdas all
      cast through size_t.

  source/backend/cpu/compute/DenseConvolutionTiledExecutor.cpp
    * hw4Stride promoted to size_t. `t_ic * hw4Stride` overflowed at
      icC4 >= 11 on full-resolution decoder layers (info[1] = 12.5M).

  source/backend/cpu/compute/ConvolutionPackWinograd.cpp
    * `sourceZStep`, `dstZStep`, and the srcStartY base-pointer
      arithmetic all promoted to size_t / cast through size_t. For
      iw*ih*batch*pack ~ 100M (full-spatial decoder Conv with batch
      = depth-as-batch = 256), `z * sourceZStep * bytes` overflows at
      z = 6.

  source/backend/cpu/x86_x64/avx/GemmCommon.cpp
    * `eReal` (= info[1] = iw*ih*batch) declared int64_t instead of
      int in both `_AVX_MNNPackC4ForMatMul_A` and the EShort variant.
      The downstream `x * unit * eReal` overflowed at x=23 / unit=8
      / eReal=12.5M (= 2.3 GB). Same int64_t treatment in the AVX
      EShort packing path.

  source/geometry/GeometryDet.cpp
  source/geometry/GeometryGather.cpp
    * Two cleanup follow-ups required by the elementSize() return-type
      change in Tensor.hpp. `auto x = output->elementSize(),
      y = input->length(...);` no longer deduces a single type
      (size_t vs int); local explicit casts.

Verified against MNN master HEAD 39be21c5 on Linux x86_64
(AMD Threadripper 3990X, NVIDIA OpenCL ICD on A100):

  * Full SIAM v0.3 fold_0 inference at native (1, 1, 256, 256, 192):
      - argmax(MNN) vs argmax(ORT-CPU) per-voxel agreement: 99.67 %
        on a real T1 (siamize tests/sub-01_T1w.nii.gz)
      - mean foreground Dice (17 classes): 0.992
      - MNN-vs-PyTorch-ref Dice (0.928) within 0.001 of ORT-vs-ref
        Dice (0.929) -- MNN adds no extra error beyond what ORT
        already adds vs the PyTorch baseline.
  * MNN-OpenCL backend works without further patches: correctness
    matches MNN-CPU, GPU sm-util hits 100 % during inference.
  * No regression on the rest of the test suite (only Conv3D / 3D-
    segmentation paths were touched).

Known remaining issues (for follow-up):

  * CPU numThread >= 4 segfaults inside CPURaster::onResize::lambda#4
    on SIAM-scale workloads. Defensive 16 MB stride threshold in this
    commit reduces the attack surface but does not fully eliminate
    the crash. Single-thread + numThread=2 both produce correct
    output; we recommend pinning numThread<=2 for 3D models with
    >2GB intermediates until this is rooted out.
  * MNN-OpenCL on SIAM v0.3 is functionally correct but ~6x slower
    than ORT-CUDA on the same A100 because the Conv3DTurn2D rewrite
    produces a 3382-op decomposed graph and OpenCL kernel-launch
    overhead dominates wall time. Not a regression -- adding a native
    Conv3D OpenCL kernel is the upstream feature that would close the
    gap.

Reproducer + investigation harness:
  https://github.com/NeuroJSON/siamize/tree/mnn/tools/mnn_probe

+++ source/backend/cpu/CPULayerNorm.cpp
@@ -86,11 +86,15 @@ ErrorCode CPULayerNorm::onExecute(const std::vector<Tensor*> &inputs,
 
     MNN_CONCURRENCY_BEGIN(ttId, threadNumber) {
         for (int tId=ttId; tId < mOutterSize; tId += threadNumber) {
-            const float* inner_input = (const float*)(input + tId * mInnerSize * bytes);
-            float* inner_output = (float*)(output + tId * mInnerSize * bytes);
+            // For 3D-segmentation models (e.g. InstanceNorm rewritten to
+            // LayerNorm over depth*height*width), mInnerSize can reach
+            // ~12.5M and tId * mInnerSize * bytes overflows int32 around
+            // tId=43. Use size_t for the byte offset.
+            const float* inner_input = (const float*)(input + (size_t)tId * mInnerSize * bytes);
+            float* inner_output = (float*)(output + (size_t)tId * mInnerSize * bytes);
             if (bytes != 4) {
-                auto tmpInput = (float*)(mTmpInputFloat.ptr() + ttId * mInnerSize * sizeof(float));
-                auto tmpOutput = (float*)(mTmpOutputFloat.ptr() + ttId * mInnerSize * sizeof(float));
+                auto tmpInput = (float*)(mTmpInputFloat.ptr() + (size_t)ttId * mInnerSize * sizeof(float));
+                auto tmpOutput = (float*)(mTmpOutputFloat.ptr() + (size_t)ttId * mInnerSize * sizeof(float));
                 if (bytes == 1) {
                     CPUCastCreator::cast(inner_input, tmpInput, CPUCastCreator::INT8_TO_FlOAT, mInnerSize, inputQuan->scale, inputQuan->zero, inputQuan->min, inputQuan->max, bn);
                 } else {
@@ -144,8 +148,8 @@ ErrorCode CPULayerNorm::onResize(const std::vector<Tensor*> &inputs,
     auto buf = bn->getBufferAllocator();
 
     if (CPUBackend::getDataType(inputs[0]) == DataType_DT_INT8 || inputs[0]->getType().bytes() == 1 || bn->functions()->bytes != 4) {
-        mTmpInputFloat = buf->alloc(threadNumber * mInnerSize * sizeof(float));
-        mTmpOutputFloat = buf->alloc(threadNumber * mInnerSize * sizeof(float));
+        mTmpInputFloat = buf->alloc((size_t)threadNumber * mInnerSize * sizeof(float));
+        mTmpOutputFloat = buf->alloc((size_t)threadNumber * mInnerSize * sizeof(float));
         buf->free(mTmpInputFloat);
         buf->free(mTmpOutputFloat);
     }
